Allow specifying specific game
This works by downloading the files.js for every patch that isn't
blacklisted and checking if any file exists that matches the game ID you
specify. Because download_single_file from thcrap_update doesn't have an
equivalent function to download to RAM, this makes the
Downloading patchlist...
part of the roulette 3x slower.